feat: use data from statsManager to render
