+++ src/plugin.ts
@@ -14,9 +14,10 @@ import { ThemeUpdater } from './themeManifest';
 import { monkeyPatchConsole } from './obsidian-hack/obsidian-mobile-debug';
 import { CalloutModal } from './callout';
 import { RecordPreview, RECORD_PREVIEW_TYPE } from './preview';
-import { STAT_PREVIEW_TYPE, Stat } from './stat'
-import StatsManager from './stats/StatsManager'
-import { pluginField, statusBarEditorPlugin, sectionWordCountEditorPlugin } from './stats/EditorPlugin'
+import { STAT_PREVIEW_TYPE, Stat } from './statsView'
+import StatsManager from './stats/statsManager'
+import { pluginField, statusBarEditorPlugin, sectionWordCountEditorPlugin } from './stats/editorPlugin'
+import { STATS_FILE_NAME } from './constant'
 
 const debug = (debug: boolean, ...msg: any) => { // eslint-disable-line @typescript-eslint/no-explicit-any
     if (debug) console.log(...msg);
@@ -88,7 +89,7 @@ export class PluginManager extends Plugin {
                 RECORD_PREVIEW_TYPE,
                 (leaf) => { return new RecordPreview(this.app, this, leaf); }
             );
-            const staticsDataDir = this.app.vault.configDir + "/vault-stats.json";
+            const staticsDataDir = this.app.vault.configDir + "/" + STATS_FILE_NAME;
             const staticsFileData = await this.app.vault.adapter.read(staticsDataDir);
             this.registerView(
                 STAT_PREVIEW_TYPE,
